backend/requirements.txt
Django>=5.0.0,<5.1.0
djangorestframework>=3.14.0,<4.0.0
django-cors-headers>=4.3.0,<5.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
gunicorn>=21.0.0,<22.0.0
